BFF-632 Moved app,bootstrap and injection file to newly created folder
